Restructuring codebase
