+++ src/Exceptions/RConException.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace DSG\SquadRCON;
+namespace DSG\SquadRCON\Exceptions;
 
 class RConException extends \Exception
 {
